add attendance

+++ onless/sign/views.py
@@ -1,5 +1,6 @@
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.db.models import Q
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render, get_object_or_404
 from django.urls import reverse_lazy
@@ -26,20 +27,28 @@ def sign(request):
 
 @login_required
 def add_schedule(request):
+    groups = Group.objects.filter(Q(is_active=True) & Q(school=request.user.school))
+    print(groups)
+    context = {
+        'groups': groups
+    }
     if request.POST:
+        get_date = request.POST.get('date').split('.')
+        date = f'{get_date[2]}-{get_date[1]}-{get_date[0]}'
         form = AddScheduleFrom(request.POST)
         author = User.objects.get(id=request.user.id)
         if form.is_valid():
             form = form.save(commit=False)
             form.author = author
+            form.date = date
+            form.subject_id = request.POST.get('subject')
             form.start = request.POST.get('start')
             form.stop = request.POST.get('stop')
-            form.pub_date = timezone.now()
             form.save()
             messages.success(request, "Muvaffaqiyatli qo'shildi !")
         else:
             messages.error(request, "Formani to'ldirishda xatolik !")
-    return render(request, 'sign/add_schedule.html')
+    return render(request, 'sign/add_schedule.html', context)
 
 
 @login_required
@@ -48,37 +57,58 @@ def update_schedule(request, id):
     context = {
         'schedule': schedule
     }
-    if request.POST:
-        form = UpdateScheduleForm(request.POST, instance=schedule)
-        if form.is_valid():
-            form = form.save(commit=False)
-            form.subject = schedule.subject
-            form.save()
-            messages.success(request, "Muvaffaqiyatli tahrirlandi !")
-            return render(request, 'sign/schedules_list.html', context)
+    if schedule.author.school == request.user.school:
+        if request.POST:
+            get_date = request.POST.get('date').split('.')
+            date = f'{get_date[2]}-{get_date[1]}-{get_date[0]}'
+            form = UpdateScheduleForm(request.POST, instance=schedule)
+            if form.is_valid():
+                form = form.save(commit=False)
+                form.date = date
+                form.start = request.POST.get('start')
+                form.stop = request.POST.get('stop')
+                form.updated_date = timezone.now()
+                form.schedule = schedule
+                form.save()
+                messages.success(request, "Muvaffaqiyatli tahrirlandi !")
+                return render(request, 'sign/schedules_list.html', context)
+            else:
+                messages.error(request, "Formani to'ldirishda xatolik !")
         else:
-            messages.error(request, "Formani to'ldirishda xatolik !")
-    else:
+            return render(request, 'sign/update_schedule.html', context)
         return render(request, 'sign/update_schedule.html', context)
-    return render(request, 'sign/update_schedule.html', context)
+    else:
+        return render(request, 'inc/404.html')
 
 
 @login_required
 def delete_schedule(request, id):
     schedule = Schedule.objects.get(id=id)
-    schedule.delete()
-    return render(request, 'sign/schedules_list.html',)
+    if schedule.author.school == request.user.school:
+        schedule.delete()
+        return render(request, 'sign/schedules_list.html',)
+    else:
+        return render(request, 'inc/404.html')
 
 
 @login_required
 def schedules_list(request):
     return render(request, 'sign/schedules_list.html')
 
+@login_required
+def subjects_list(request):
+    subjects = Subject.objects.filter(is_active=True)
+    context = {
+        'subjects': subjects
+    }
+    return render(request, 'sign/subjects_list.html', context)
 
 @login_required
 def get_subject(request):
     if request.is_ajax():
-        subjects = Subject.objects.filter(category=request.GET.get('category'))
+        group = get_object_or_404(Group, id=request.GET.get('group'))
+        subjects = Subject.objects.filter(category=group.category)
+        print(subjects)
         options = "<option>-- -- --</option>"
         for subject in subjects:
             options += f"<option value='{subject.id}'>{subject.title}</option>"
@@ -99,10 +129,11 @@ def get_schedule(request):
 
             td += f"<tr>" \
                   f"<td>M-{schedule.sort}</td>" \
-                  f"<td>{schedule.title}</td>" \
+                  f"<td class='schedule_title'>{schedule.title}</td>" \
+                  f"<td>{schedule.date}</td>" \
                   f"<td>{str(schedule.start)[0:16]}</td>" \
                   f"<td>{str(schedule.stop)[0:16]}</td>" \
-                  f"<td><a href='{url_update}'><i class='fa fa-edit'></i></a><a href='{url_delete}'><i class='fa fa-trash ml-2'></i></a></td>" \
+                  f"<td class='edit_block'><a href='{url_update}'><i class='fa fa-edit'></i></a><a class='deleteBtn' data-title='{schedule.title}' data-id='{schedule.id}' href='#'><i class='fa fa-trash ml-2'></i></a></td>" \
                   f"</tr>"
         return HttpResponse(td)
     else:
